platforms: make args consistent for NuttX/Posix
The implementation of task_spawn in NuttX is according to the POSIX
specs and adds the name of the task as the first argument `argv[0]`.
For Linux/macOS/Cygwin we did not adhere to this in
`px4_task_spawn_cmd`. This meant that some of the modules manually did

`#ifdef __PX4_NUTTX argc -= 1; argv += 1; #endif`

to overcome the inconsistency.

This change makes the "posix" implementation behave like NuttX.

Also, in this commit, we remove the NutX only hacks to remove the
first argument in various places because it is now needed generally.